requirements.txt
fastapi
uvicorn
groq
mem0ai
qdrant-client
pydantic-settings
python-dotenv
firecrawl-py
fastembed
python-multipart
